Coding Exercise - Straker Translations
Aim:

The aim of this exercise is to use Vue.js. It requires that you write a single page app that retrieves data from the API. When the app loads, it should retrieve the list of users from API and display them. Clicking on the name of the user should display 10 recent posts written by that user. Clicking on the title of the post displays the details of the post.